Move sys.exit calls to script logic
This will prevent unintended exits when the code is consumed as a
library. The functionality is preserved for script users.

Tests were updated to check for the exception type.

Additional change includes closing the `bin_file` when an Exception is
thrown. There were warnings in the test output about this.